Refactor read_chunk, work on HlsStream.
