+++ rust/src/audio/streaming/streamable.rs
@@ -24,7 +24,20 @@ pub const CHUNK_SIZE:usize = 1024 * 128;
 
 pub trait Streamable<T: Read + Seek + Send + Sync + MediaSource>
 {
-    fn read_chunk(tx:Sender<(usize, Vec<u8>)>, url:String, start:usize, file_size:usize);
+    /// Gets the next chunk in the sequence.
+    /// 
+    /// Returns the received bytes by sending them via `tx`.
+    fn read_chunk(tx:Sender<(usize, Vec<u8>)>, url:String, start:usize, file_size:usize)
+    {
+        let end = (start + CHUNK_SIZE).min(file_size) - 1;
+
+        let chunk = reqwest::blocking::Client::new().get(url)
+            .header("Range", format!("bytes={start}-{end}"))
+            .send().unwrap().bytes().unwrap().to_vec();
+        
+        tx.send((start, chunk)).unwrap();
+    }
+
     fn try_write_chunk(&mut self, should_buffer:bool);
     fn should_get_chunk(&self) -> (bool, usize);
 }
